pyproject.toml
[project]
name = "default-fast-api-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "alembic>=1.17.2",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.125.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
